Build on windows (#3)
* Windows build
* Update blibs for MSVC
* windows: Add missing files
* rom2exe: Add .exe extension for windows
* windows: Add CI
* Suppress clangd's unused warnings

+++ src/asm.c
@@ -676,7 +676,7 @@ main(int argc, const char* argv[]) {
 			buxn_dbg_symtab_io_status_t status = buxn_dbg_write_symtab(
 				writer,
 				&(buxn_dbg_symtab_t){
-					.num_symbols = barray_len(ctx.debug_symbols),
+					.num_symbols = (uint32_t)barray_len(ctx.debug_symbols),
 					.symbols = ctx.debug_symbols,
 				}
 			);
